MDMP v1.1 Tests (#629)
* Moved the mdmp tests to extra

Moved the mdmp tests to the extras section and have started to create
valid regression tests.

* Readded the entrypoint test.

This is a test that is known to fail and will/should probably be fixed
later on in the plugins development.

* Corrected file name convention

Corrected the dump file extension to .dmp.

* Added the missing calc.dmp file

Added the file that was suposedly missing.

* Changed the extension in the mdmp file

Forgot to change the dmp extension in the mdmp file.

* Added the mdmp tests to the Makefile

Added the mdmp test to the Makefile under the name extras.mdmp.

* Added entrypoint regression tests

Added tests to handle the new entrypoint functionality.

* Additional MDMP tests

Added additional MDMP tests for the new set of functionality that has
been added to the plugin.

* New tests

Added better hello dumps that have actually finished loading. This
allows us to perform tests like checking that relocs point to valid
areas!

* Tests for MDMP v1.1

Added/fixed tests to work with MDMP v1.1.